# update first byte timeout algo